+++ src/main/java/org/example/entities/hostileEntity/enemy/Pterodactyl.java
@@ -3,7 +3,7 @@ package org.example.entities.hostileEntity.enemy;
 import com.github.hanyaeger.api.Coordinate2D;
 
 public class Pterodactyl extends Enemy {
-    public Pterodactyl(Coordinate2D initialLocation, int speed) {
+    public Pterodactyl(Coordinate2D initialLocation, double speed) {
         super("sprites/zwaard.png", initialLocation, speed);
     }
 }
