+++ src/main/java/org/example/entities/hostileEntity/obstacle/Obstacle.java
@@ -4,7 +4,7 @@ import com.github.hanyaeger.api.Coordinate2D;
 import org.example.entities.hostileEntity.HostileEntity;
 
 public abstract class Obstacle extends HostileEntity {
-    protected Obstacle(String resource, Coordinate2D initialLocation, int speed) {
+    protected Obstacle(String resource, Coordinate2D initialLocation, double speed) {
         super(resource, initialLocation, speed);
     }
 }
