+++ src/main/java/org/example/entities/hostileEntity/obstacle/Trashcan.java
@@ -3,7 +3,7 @@ package org.example.entities.hostileEntity.obstacle;
 import com.github.hanyaeger.api.Coordinate2D;
 
 public class Trashcan extends Obstacle {
-    public Trashcan(Coordinate2D initialLocation, int speed) {
+    public Trashcan(Coordinate2D initialLocation, double speed) {
         super("sprites/prullenbak.png", initialLocation, speed);
     }
 }
